refactor: cut video exporter registry — replace virtual dispatch with direct if/else
Delete 7 dead files (codex/cut-video-exporter-registry):

- exporter_registry.cpp/.hpp: unordered_map<string, unique_ptr<VideoExporter>>
  lookup — a glorified switch statement over 4 exporters.
- exporter.hpp: VideoExporter virtual base class + VideoExportJob struct.
- pipe_exporter.cpp: 19-line wrapper calling render_and_encode_ffmpeg_pipe.
- chunked_exporter.cpp: 18-line wrapper calling render_and_encode_ffmpeg_chunked.
- null_render_exporter.cpp: render loop benchmark duplicate of bench command.
- null_convert_exporter.cpp: convert benchmark duplicate of dev bench-convert.

Replace with direct dispatch:
- render_and_encode_ffmpeg() now does a simple if/else (pipe/png) calling
  render_and_encode_ffmpeg_pipe or render_and_encode_ffmpeg_chunked directly.
- execute_video_job() rejects non-FFmpeg sinks with error message.
- validate_video_job() drops registry dependency — just validates mode string.
- video_job_plan.cpp: simplified sink type parsing (always Ffmpeg).
- register_video_commands.cpp: removed --sink flag (null-render/null-convert).
- video_job_plan.hpp: removed ExporterRegistry forward declaration.

Also fix pre-existing narrowing conversion in pipe_export_pool_warmup.cpp
(TextSizeClass::count int→size_t) and stale PipeExporter comment.

+++ apps/chronon3d_cli/commands/video/exporters/pipe_export_pool_warmup.cpp
@@ -22,7 +22,7 @@ namespace {
 // margins). Pre-warming them gives exact-hit + best-fit reuse instead of
 // fresh allocations on the hot EffectStack path.
 void warmup_text_size_classes(cache::FramebufferPool& pool) {
-    struct TextSizeClass { int w; int h; int count; const char* label; };
+    struct TextSizeClass { int w; int h; size_t count; const char* label; };
     // Counts tuned against the FramebufferPool default budget (384 MB) so the
     // total preallocation stays well under the cap. Color is float4 = 16 B/px,
     // so e.g. the canvas bucket (1920×1152) costs ~35 MB per buffer. The
